+++ patches/Accessibility/strings_pl.cpp
@@ -586,12 +586,15 @@ const char* Get(Id id) {
 
         case Id::HelpKeyModSettings:     return "Ustawienia moda s\xB9 w Opcjach, na dole listy";
         case Id::HelpKeyPadMenuNav:          return "W menu: krzy\xBF" "ak lub lewa ga\xB3ka przesuwaj\xB9, A zatwierdza, B cofa";
+        case Id::HelpKeyPadTooltip:          return "Przytrzymaj Y i naci\x9Cnij krzy\xBF" "ak w g\xF3r\xEA lub w d\xF3\xB3: przeczytaj ca\xB3y opis bez zmiany pozycji";
         case Id::HelpKeyPadInteract:          return "A: domy\x9Clna akcja na wybranym celu (atak, otwarcie, rozmowa, podniesienie)";
         case Id::HelpKeyPadCycleObjects:     return "Krzy\xBF" "ak w lewo i w prawo: poprzedni i nast\xEApny obiekt";
         case Id::HelpKeyPadCycleCategory:    return "Krzy\xBF" "ak w g\xF3r\xEA i w d\xF3\xB3: poprzednia i nast\xEApna kategoria";
         case Id::HelpKeyPadMapHints:         return "Ekran mapy, krzy\xBF" "ak w lewo i w prawo: poprzedni i nast\xEApny znacznik mapy";
         case Id::HelpKeyPadActionMenuToggle: return "Lewy spust: otwiera menu akcji, ponowne naci\x9Cni\xEA" "cie je zamyka";
         case Id::HelpKeyPadActionMenuNav:    return "Menu akcji, krzy\xBF" "ak: w lewo i w prawo zmienia kategori\xEA, w g\xF3r\xEA i w d\xF3\xB3 pozycj\xEA, A wykonuje, B lub lewy spust zamyka";
+        case Id::HelpKeyPadSelfStatus:       return "Lewy spust z X: tw\xF3j stan";
+        case Id::HelpKeyPadActionQueue:      return "Prawy spust z X: kolejka akcji. Krzy\xBF" "ak w g\xF3r\xEA i w d\xF3\xB3 przechodzi po pozycjach, A usuwa ostatni\xB9 w kolejce, lewy spust z A czy\x9C" "ci ca\xB3\xB9 kolejk\xEA, B zamyka";
         case Id::HelpKeyPadWalkToFocus:      return "Prawy spust z prawym bumperem: id\x9F do wybranego obiektu";
         case Id::HelpKeyPadBeacon:           return "Lewy spust z lewym bumperem: sygna\xB3 d\x9Fwi\xEAkowy do wybranego obiektu";
         case Id::HelpKeyPadDegrees:          return "Prawy spust: podaj kierunek w stopniach";
@@ -601,9 +604,9 @@ const char* Get(Id id) {
         case Id::HelpKeyPadCycleTargets:     return "Lewy i prawy bumper: zmiana celu";
         case Id::HelpKeyPadTradeMode:        return "Pojemnik i sklep, bumpery: zmiana trybu";
         case Id::HelpKeyPadOptions:          return "Przycisk Back: menu opcji";
+        case Id::HelpKeyPadPause:            return "Start: pauza";
         case Id::HelpKeyPadSwitchLeader:
             return "X: zmiana przyw\xF3" "dcy dru\xBFyny";
-        case Id::HelpKeyPadPause:            return "Start: pauza";
         case Id::PadQuickMenuOpened:
             return "Szybkie menu. G\xF3ra i d\xF3\xB3 wybiera, A zatwierdza, "
                    "B zamyka.";